requirements.txt
Flask
pytesseract
Pillow
requests
gunicorn
shelve
